+++ warthog/__init__.py
@@ -13,4 +13,4 @@ warthog
 
 """
 
-__version__ = '0.5.0'
+__version__ = '0.6.0'